requirements.txt
Flask
WeasyPrint
Flask-SQLAlchemy
